# fix trimming of right side