--- a/src/Buffer.test.ts
+++ b/src/Buffer.test.ts
@@ -508,11 +508,10 @@ describe('Buffer', () => {
       // the dangling last cell is wrongly added in the string
       // --> fixable after resolving #1685
       terminal.writeSync(input);
-      // TODO: reenable after fix
       const s = terminal.buffer.iterator(true).next().content;
       assert.equal(input, s);
       for (let i = 10; i < input.length; ++i) {
-        const bufferIndex = terminal.buffer.stringIndexToBufferIndex(0, i); // TODO: remove +1 after fix
+        const bufferIndex = terminal.buffer.stringIndexToBufferIndex(0, i);
         const j = (i - 0) << 1;
         assert.deepEqual([(j / terminal.cols) | 0, j % terminal.cols], bufferIndex);
       }