@jerch 2018-11-08
Up for next review. Note that this also needs user testing before it can be added, since it is used at many places (not sure about the test coverage).